Split public api from feeds client
