+++ packages/stream_feeds/lib/src/repositories.dart
@@ -1,1 +1,2 @@
 export 'repositories/feeds_repository.dart';
+export 'repositories/repository_provider.dart';
